requirements.txt
apache-airflow-providers-apache-spark
pyspark==4.0.0
